split into instant casts and non instant casts.
