+++ src/Parser/Core/Modules/EarlyDotRefreshes/EarlyDotRefreshesInstantsSuggestion.js
